Moved codegen::If + codegen::Loop into 'codegen/utils' #690
