+++ src/codegen/hash_join_translator.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/hash_join_translator.h"
 
-#include "codegen/if.h"
+#include "include/codegen/utils/if.h"
 #include "codegen/oa_hash_table_proxy.h"
 #include "include/codegen/expression/tuple_value_translator.h"
 #include "codegen/vectorized_loop.h"
