+++ src/codegen/oa_hash_table.cpp
@@ -13,8 +13,8 @@
 #include "codegen/oa_hash_table.h"
 
 #include "codegen/hash.h"
-#include "codegen/if.h"
-#include "codegen/loop.h"
+#include "include/codegen/utils/if.h"
+#include "include/codegen/utils/loop.h"
 #include "codegen/oa_hash_table_proxy.h"
 #include "codegen/vectorized_loop.h"
 #include "codegen/utils/oa_hash_table.h"
