+++ src/codegen/table.cpp
@@ -14,7 +14,7 @@
 
 #include "catalog/schema.h"
 #include "codegen/data_table_proxy.h"
-#include "codegen/loop.h"
+#include "include/codegen/utils/loop.h"
 #include "codegen/runtime_functions_proxy.h"
 #include "storage/data_table.h"
