+++ src/codegen/utils/if.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/if.h"
+#include "include/codegen/utils/if.h"
 
 #include "llvm/Transforms/Utils/BasicBlockUtils.h"
