+++ src/codegen/utils/loop.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/loop.h"
+#include "include/codegen/utils/loop.h"
 
 namespace peloton {
 namespace codegen {
