+++ src/codegen/value.cpp
@@ -15,7 +15,7 @@
 #include <list>
 #include <queue>
 
-#include "codegen/if.h"
+#include "include/codegen/utils/if.h"
 #include "codegen/runtime_functions_proxy.h"
 #include "codegen/type.h"
