+++ src/include/codegen/varlen.h
@@ -13,7 +13,7 @@
 #pragma once
 
 #include "codegen/codegen.h"
-#include "codegen/if.h"
+#include "include/codegen/utils/if.h"
 #include "codegen/runtime_functions_proxy.h"
 #include "codegen/varlen_proxy.h"
